streamlit_app/requirements.txt
oandapyV20==0.7.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
matplotlib==3.7.2
tensorflow==2.15.0
xgboost==2.0.3
seaborn==0.13.1
